refactored project structre and built custom option selector for contacts
