+++ src/components/PluginsTab/AddNewPlugin/AddNewPlugin.tsx
