+++ src/components/PluginsTab/AddNewPlugin/PrimaIntegrationBtns.tsx
