+++ src/components/PluginsTab/AddNewPlugin/ShowPlugin.tsx
