+++ src/components/PluginsTab/MainView.tsx
