+++ src/theme/theme.tsx
@@ -130,16 +130,6 @@ const theme = extendTheme({
       bg: "darkGray",
     },
   },
-
-  CheckBox: {
-    ".chakra-checkbox__control": {
-      border: "3px solid #5BA4AF",
-      padding: "10px",
-    },
-    ".chakra-checkbox__label": {
-      marginInlineStart: "2.5em",
-    },
-  },
 });
 
 export default theme;
